+++ src/main/kotlin/net/freshplatform/ktor_server/firebase_app_check/utils/extensions/RouteExtensions.kt
@@ -1,4 +1,0 @@
-package net.freshplatform.ktor_server.firebase_app_check.utils
-
-class RouteExtensions {
-}